optimize block vertex

+++ src/chunk.hpp
@@ -4,6 +4,7 @@
 #include "block.hpp"
 #include "block_vertex_array.hpp"
 #include <glm/glm.hpp>
+#include <sys/types.h>
 
 namespace kwnc
 {
@@ -33,8 +34,8 @@ private:
         Block *blocks = nullptr;
         bool dirty = false;
 
-        static int convert_to_block_idx(int x, int y, int z);
-        static void convert_to_pos_in_chunk(int i, int *x, int *y, int *z);
+        static int convert_to_block_idx(u_int8_t x, u_int8_t y, u_int8_t z);
+        static void convert_to_pos_in_chunk(int i, u_int8_t *x, u_int8_t *y, u_int8_t *z);
 };
 } // namespace kwnc
